Time: 803 ms (34.50%), Space: 227.6 MB (16.15%) - LeetHub

+++ 2477-minimum-fuel-cost-to-report-to-the-capital/2477-minimum-fuel-cost-to-report-to-the-capital.cpp
@@ -1,5 +1,5 @@
 class Solution {
-        unordered_map<int,vector<int>>m;
+    unordered_map<int,vector<int>>m;
     long long ans;
     int seat;
     vector<int>vis;
@@ -16,12 +16,12 @@ public:
         dfs(0);
         return ans;
     }
-    long long dfs(int node){
+    int dfs(int node){
         vis[node]=1;
-        long long res=1;
+        int res=1;
         for(auto child:m[node]){
             if(vis[child]==0){
-                long long temp=dfs(child);
+                int temp=dfs(child);
                 ans+=(temp+seat-1)/seat;  
                 res+=temp;
             }
